requirements.txt
absl-py==2.0.0
annotated-types==0.5.0
anyio==3.7.1
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cycler==0.11.0
dnspython==2.4.2
email-validator==2.0.0.post2
ember @ git+https://github.com/elastic/ember.git@d97a0b523de02f3fe5ea6089d080abacab6ee931
environs==9.5.0
fastapi==0.103.1
fastapi_cors==0.0.6
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
lief==0.13.2
lightgbm==4.1.0
Markdown==3.4.4
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.9.7
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
protobuf==4.24.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scapy==2.5.0
scikit-learn==1.3.0
scipy==1.11.2
sh==2.0.6
six==1.16.0
sniffio==1.3.0
sse-starlette==1.6.5
starlette==0.27.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchdog==3.0.0
watchfiles==0.20.0
websockets==11.0.3
Werkzeug==2.3.7
wordcloud==1.9.2
wrapt==1.15.0
xgboost==2.0.0
yara-python==4.3.1
